delete feito

+++ semana17/aula1/aula-back-arquitetura-exercicio-tarde-inicio/src/business/usecase/deleteUser.ts
@@ -1,17 +1,22 @@
-import {UserDB} from '../../data/userDataBase'
+import { UserDB } from "../../data/userDataBase";
 
-export class DeleteUserUC{
-    constructor(private userDB: UserDB){}
+export class DeleteUserUC {
+  constructor(private userDB: UserDB) {}
 
-    public async execute(input: DeleteUserByIdUCInput): Promise<void>{
-        const user = await this.userDB.deleteUser(input.id)
-
-        if(!user){
-            throw new Error("user not found")
-        }
+  public async execute(input: DeleteUserByIdUCInput): Promise<void> {
+    if (!input.id) {
+      throw new Error("user id is invalid");
     }
+
+    const user = await this.userDB.getUserById(input.id);
+    if (!user) {
+      throw new Error("User not found");
+    }
+
+    await this.userDB.deleteUser(input.id);
+  }
 }
 
 export interface DeleteUserByIdUCInput {
-    id: string;
-  }
+  id: string;
+}
